backend/elearning_project/pyproject.toml
[tool.poetry]
name = "elearning-project"
version = "0.1.0"
description = ""
authors = ["anujkarn002 <anujkarn002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.4.0"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework-camel-case = "^1.4.2"
django-extensions = "^3.2.3"
channels = "^4.1.0"
django-storages = "^1.14.4"
drf-nested-routers = "^0.94.1"
drf-spectacular = "^0.27.2"
redis = "^5.0.8"
model-mommy = "^2.0.0"
flake8 = "^7.1.1"
envparse = "^0.2.0"
django-filter = "^24.3"
django-celery-beat = "^2.7.0"
django-seeding = "^1.3.0"
pandas = "^2.2.2"
daphne = "^4.1.2"
pyparsing = "^3.1.4"
pydot = "^3.0.1"


[tool.isort]
profile = "black"
src_paths = ["elearning_project",]
combine_as_imports = true

[tool.pytest.ini_options]
testpaths = "tests"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::RuntimeWarning",
    "ignore::UserWarning",
]
addopts = '''
  --strict-markers
  --tb=short
  --cov=app
  --cov=tests
  --cov-branch
  --cov-report=term-missing
  --cov-report=html
  --cov-report=xml
  --no-cov-on-fail
  --cov-fail-under=30
  --numprocesses=auto
  --asyncio-mode=auto
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
